requirements.txt
aiogram==2.22.1
aiohttp==3.8.3
aiosignal==1.2.0
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.0
frozenlist==1.3.1
idna==3.4
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
matplotlib-inline==0.1.6
multidict==6.0.2
nest-asyncio==1.5.6
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pyzmq==24.0.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.5.1
toml==0.10.2
tornado==6.2
traitlets==5.4.0
urllib3==1.26.12
wcwidth==0.2.5
yarl==1.8.1
zodiac-sign==0.2.5
